Adds an instantiable check for all classes in the class map.
The main goal of this change is to remove all invalid class form
the class map that are not able to return by a factory. This means
interfaces, abstract classes, traits, class with private
constructors, class with invalid class name and broken classes that
can't define. To use the additional check the cli parameter
--instantiableCheck should use.
With the new debug flag, it is possible to see, why a class gets
excluded from the class map.
It also possible to extend the class map without changing the code by
 using the event "phpstorm_map_generator_extend_map".